update miraheze table automatically
Adjust existing scripts that get regular miraheze
wikis and miraheze wikis with custom domains and
combine them in a third script that can be called via cron
to automatically update the entire table and resolve T153930.

Bug: T153930

+++ usr/local/bin/wikistats/import_miraheze_custom_wikis.sh
@@ -9,7 +9,14 @@ OUTFILE="${WORKDIR}/miraheze_custom_wikis.sql"
 
 
 cd $WORKDIR
-/bin/rm $INFILE
+
+if [ -f $INFILE ]; then
+    # echo -e "$INFILE exists. deleting to clean up\n"
+    /bin/rm $INFILE
+fi
+
+# echo -e "wget'ting $URL\n"
+
 /usr/bin/wget $URL
 
 while read -r line
@@ -18,14 +25,17 @@ while read -r line
     wiki=$(echo $line|cut -d\| -f1)
     url=$(echo $line| cut -d\| -f2)
 
-    echo "DELETE from miraheze where prefix='${wiki}';" | tee $OUTFILE
+    # echo "DELETE from miraheze where prefix='${wiki}';" | tee $OUTFILE
     echo "INSERT IGNORE INTO miraheze (prefix,method,statsurl) values ('${wiki}','8','${url}');" | tee $OUTFILE
 
 done < "${INFILE}"
 
 cat $OUTFILE
 
-/usr/bin/mysql -u root wikistats < $OUTFILE
+# echo -e "importing to mysql ..\n"
+# /usr/bin/mysql -u root wikistats < $OUTFILE
 
+# echo -e "cleaning up. deleting work files. bye\n"
 /bin/rm $WORKDIR/$INFILE $OUTFILE
 
+exit
